pyproject.toml
[project]
name = "bafoka-voice-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.122.0",
    "httpx>=0.28.1",
    "numpy>=2.3.5",
    "pydantic-settings>=2.12.0",
    "torch>=2.9.1",
    "whisper>=1.1.10",
]
